+++ packages/angular-workspace/nimble-angular/label-provider/core/nimble-label-provider-core.directive.ts
@@ -91,7 +91,7 @@ export class NimbleLabelProviderCoreDirective {
         return this.elementRef.nativeElement.scrollBackward;
     }
 
-    @Input('scrollBackward') public set scrollBackward(value: string | undefined) {
+    @Input('scroll-backward') public set scrollBackward(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'scrollBackward', value);
     }
 
@@ -99,7 +99,15 @@ export class NimbleLabelProviderCoreDirective {
         return this.elementRef.nativeElement.scrollForward;
     }
 
-    @Input('scrollForward') public set scrollForward(value: string | undefined) {
+    @Input('scroll-forward') public set scrollForward(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'scrollForward', value);
+    }
+
+    public get itemRemove(): string | undefined {
+        return this.elementRef.nativeElement.itemRemove;
+    }
+
+    @Input('item-remove') public set itemRemove(value: string | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'itemRemove', value);
     }
 }