+++ packages/spright-components/src/all-components.ts
@@ -9,4 +9,7 @@ import '@ni/nimble-components/dist/esm/all-components';
 import './chat/conversation';
 import './chat/input';
 import './chat/message';
+import './chat/message/inbound';
+import './chat/message/outbound';
+import './chat/message/system';
 import './rectangle';